fixed identity registration test
