+++ Cargo.toml
@@ -15,7 +15,7 @@ repository = "https://github.com/yourusername/soroban-project"
 
 
 [workspace.dependencies]
-soroban-sdk = "20.3.0"
+soroban-sdk = "21.0.0"
